+++ rbx_xml/src/serializer.rs
@@ -12,15 +12,17 @@ use crate::{
         serialize_binary_string,
         serialize_bool,
         serialize_cframe,
-        serialize_content,
         serialize_color3,
         serialize_color3uint8,
+        serialize_content,
         serialize_enum,
         serialize_float32,
         serialize_int32,
         serialize_physical_properties,
         serialize_ref,
         serialize_string,
+        serialize_udim,
+        serialize_udim2,
         serialize_vector2,
         serialize_vector2int16,
         serialize_vector3,
@@ -98,6 +100,12 @@ impl<W: Write> XmlEventWriter<W> {
         Ok(())
     }
 
+    pub fn write_tag_characters<T: std::fmt::Display>(&mut self, tag: &str, value: T) -> Result<(), writer::Error> {
+        self.write(XmlWriteEvent::start_element(tag))?;
+        self.write_characters(value)?;
+        self.write(XmlWriteEvent::end_element())
+    }
+
     pub fn write_tag_array<T: std::fmt::Display>(&mut self, values: &[T], tags: &[&str]) -> Result<(), writer::Error> {
         assert_eq!(values.len(), tags.len());
 
@@ -120,23 +128,27 @@ fn serialize_value<W: Write>(
         .get(canonical_name)
         .unwrap_or(&canonical_name);
 
+    #[allow(unreachable_patterns)]
     match value {
-        RbxValue::Bool { value } => serialize_bool(writer, xml_name, *value),
-        RbxValue::String { value } => serialize_string(writer, xml_name, value),
         RbxValue::BinaryString { value } => serialize_binary_string(writer, xml_name, value),
-        RbxValue::Vector2 { value } => serialize_vector2(writer, xml_name, *value),
-        RbxValue::Vector3 { value } => serialize_vector3(writer, xml_name, *value),
-        RbxValue::Vector2int16 { value } => serialize_vector2int16(writer, xml_name, *value),
-        RbxValue::Vector3int16 { value } => serialize_vector3int16(writer, xml_name, *value),
-        RbxValue::Float32 { value } => serialize_float32(writer, xml_name, *value),
-        RbxValue::Int32 { value } => serialize_int32(writer, xml_name, *value),
-        RbxValue::Enum { value } => serialize_enum(writer, xml_name, *value),
-        RbxValue::PhysicalProperties { value } => serialize_physical_properties(writer, xml_name, *value),
-        RbxValue::Ref { value } => serialize_ref(writer, xml_name, *value),
+        RbxValue::Bool { value } => serialize_bool(writer, xml_name, *value),
         RbxValue::CFrame { value } => serialize_cframe(writer, xml_name, *value),
-        RbxValue::Content { value } => serialize_content(writer, xml_name, value),
         RbxValue::Color3 { value } => serialize_color3(writer, xml_name, *value),
         RbxValue::Color3uint8 { value } => serialize_color3uint8(writer, xml_name, *value),
+        RbxValue::Content { value } => serialize_content(writer, xml_name, value),
+        RbxValue::Enum { value } => serialize_enum(writer, xml_name, *value),
+        RbxValue::Float32 { value } => serialize_float32(writer, xml_name, *value),
+        RbxValue::Int32 { value } => serialize_int32(writer, xml_name, *value),
+        RbxValue::PhysicalProperties { value } => serialize_physical_properties(writer, xml_name, *value),
+        RbxValue::Ref { value } => serialize_ref(writer, xml_name, *value),
+        RbxValue::String { value } => serialize_string(writer, xml_name, value),
+        RbxValue::UDim { value } => serialize_udim(writer, xml_name, *value),
+        RbxValue::UDim2 { value } => serialize_udim2(writer, xml_name, *value),
+        RbxValue::Vector2 { value } => serialize_vector2(writer, xml_name, *value),
+        RbxValue::Vector2int16 { value } => serialize_vector2int16(writer, xml_name, *value),
+        RbxValue::Vector3 { value } => serialize_vector3(writer, xml_name, *value),
+        RbxValue::Vector3int16 { value } => serialize_vector3int16(writer, xml_name, *value),
+
         unknown => {
             warn!("Property value {:?} cannot be serialized yet", unknown);
             unimplemented!();