+++ rbx_xml/src/types/mod.rs
@@ -1,23 +1,25 @@
 mod binary_string;
 mod bool;
 mod cframe;
-mod content;
 mod colors;
+mod content;
 mod enumeration;
 mod numbers;
 mod physical_properties;
 mod r#ref;
 mod string;
+mod udim;
 mod vectors;
 
 pub use self::binary_string::*;
 pub use self::bool::*;
 pub use self::cframe::*;
-pub use self::content::*;
 pub use self::colors::*;
+pub use self::content::*;
 pub use self::enumeration::*;
 pub use self::numbers::*;
 pub use self::physical_properties::*;
 pub use self::r#ref::*;
 pub use self::string::*;
+pub use self::udim::*;
 pub use self::vectors::*;